+++ release-notes.md
@@ -4,6 +4,8 @@ Pending changes
 Changes:
 - Implemented a new /ui/app/inbox.html route to upload and possible process files
   during upload.  Check the "Inbox" configuration (TODO: doc)
+- Added support to display audit-logs.
+
 
 1.8.5 (2025-06-17)
 ==================